Tests: Add mocks for cleaner testing on local PCs.

+++ __tests__/main.test.ts
@@ -1,3 +1,6 @@
+// Mocking OS for os.homedir()
+jest.mock('os');
+
 import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
@@ -10,26 +13,30 @@ import * as gcc from '../src/gcc';
 import * as setup from '../src/setup';
 
 const TEMP_LOCAL_PATH = path.join(__dirname, '..', 'TESTS_TEMP_DELETE');
+const TEMP_HOME_DIR = path.join(TEMP_LOCAL_PATH, 'HOME');
+const TEMP_CACHE_DIR = path.join(TEMP_LOCAL_PATH, 'CACHE');
 
 jest.setTimeout(5 * 60 * 1000);
 
 beforeAll(() => {
-  if (fs.existsSync(TEMP_LOCAL_PATH)) {
-    rimraf.sync(TEMP_LOCAL_PATH);
-  }
-  fs.mkdirSync(TEMP_LOCAL_PATH);
+  if (fs.existsSync(TEMP_HOME_DIR)) rimraf.sync(TEMP_HOME_DIR);
+  if (fs.existsSync(TEMP_CACHE_DIR)) rimraf.sync(TEMP_CACHE_DIR);
+  if (fs.existsSync(TEMP_LOCAL_PATH)) rimraf.sync(TEMP_LOCAL_PATH);
 
-  // Mocking os.homedir() result
-  const homeDir = path.join(TEMP_LOCAL_PATH, 'HOME');
-  fs.mkdirSync(homeDir);
-  if (os.platform() === 'win32') {
-    process.env['USERPROFILE'] = homeDir;
-  } else {
-    process.env['HOME'] = homeDir;
-  }
+  fs.mkdirSync(TEMP_LOCAL_PATH);
+  fs.mkdirSync(TEMP_HOME_DIR);
+  fs.mkdirSync(TEMP_CACHE_DIR);
+
+  // Mocking os.homedir() result at the top of this file
+  console.warn(`Testing home path: ${os.homedir()}`);
+
   // Env vars needed for the GitHub actions libs
   process.env['RUNNER_TEMP'] = TEMP_LOCAL_PATH;
-  process.env['RUNNER_TOOL_CACHE'] = path.join(TEMP_LOCAL_PATH, 'CACHE');
+  process.env['RUNNER_TOOL_CACHE'] = TEMP_CACHE_DIR;
+
+  // Try to reduce GH lib verbosity
+  process.env['ACTIONS_STEP_DEBUG'] = '';
+  process.env['RUNNER_DEBUG'] = '';
 });
 
 afterAll(done => {
